+++ dataframe-jdbc/src/test/kotlin/org/jetbrains/kotlinx/dataframe/io/h2/mysqlH2Test.kt
@@ -403,7 +403,7 @@ class MySqlH2Test {
         schema.columns["mediumintcol"]!!.type shouldBe typeOf<Int>()
         schema.columns["mediumintunsignedcol"]!!.type shouldBe typeOf<Int>()
         schema.columns["bigintcol"]!!.type shouldBe typeOf<Long>()
-        schema.columns["floatcol"]!!.type shouldBe typeOf<Float>()
+        schema.columns["floatcol"]!!.type shouldBe typeOf<Double>()
         schema.columns["doublecol"]!!.type shouldBe typeOf<Double>()
         schema.columns["decimalcol"]!!.type shouldBe typeOf<BigDecimal>()
     }
